Use doubles instead of integers for result values
This change had been partly completed before, however the
adding of results/training of a machine learner still used
integer values instead of floating point values. This is
important for results which may not necessarily be integer
values such as energy.

+++ include/mageec/Database.h
@@ -216,7 +216,7 @@ public:
   ///
   /// \param results A set of results to be added to the database
   void
-  addResults(std::map<std::pair<CompilationID, std::string>, uint64_t> results);
+  addResults(std::map<std::pair<CompilationID, std::string>, double> results);
 
 //===----------------------- Training interface ---------------------------===//
